requirements.txt
streamlit
numpy>=1.24.0
plotly>=5.18.0
pandas
openpyxl==3.1.2
authlib

python-dotenv==1.0.1
google-generativeai>=0.3.2
python-dateutil
pyyaml
streamlit-option-menu
fuzzywuzzy
python-Levenshtein
msal
streamlit-javascript
streamlit_lottie

supabase>=2.3.0
postgrest>=0.10.0
storage3>=0.5.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
